move mixin functions into interface
